+++ code/packages/rust/cas-pretty-printer/src/mathematica.rs
@@ -14,9 +14,7 @@
 
 use symbolic_ir::{IRApply, IRNode};
 
-use crate::dialect::{
-    default_binary_op, default_precedence, default_unary_op, Dialect,
-};
+use crate::dialect::{default_binary_op, default_precedence, default_unary_op, Dialect};
 use crate::macsyma::macsyma_sugar;
 
 // ---------------------------------------------------------------------------